requirements.txt
vk_api==11.8.0
google_api_python_client==1.8.3
python-dotenv==0.13.0
